backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pandas>=2.2.0
duckdb>=1.0.0
scikit-learn>=1.5.0
xgboost>=2.1.0
python-dotenv>=1.0.0
openai>=1.51.0
